Move RepoListActivity to ui package
