+++ BrusovCodeTest/app/src/main/AndroidManifest.xml
@@ -9,7 +9,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".RepoListActivity">
+        <activity android:name=".ui.repolist.RepoListActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
