+++ BrusovCodeTest/app/src/main/java/com/eugenebrusov/brusovcodetest/ui/repolist/RepoListActivity.kt
@@ -1,7 +1,8 @@
-package com.eugenebrusov.brusovcodetest
+package com.eugenebrusov.brusovcodetest.ui.repolist
 
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
+import com.eugenebrusov.brusovcodetest.R
 
 class RepoListActivity : AppCompatActivity() {
