+++ BrusovCodeTest/app/src/main/res/layout/activity_repo_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RepoListActivity">
+    tools:context=".ui.repolist.RepoListActivity">
 
     <TextView
         android:layout_width="wrap_content"
